pyproject.toml
[tool.poetry]
name = "testing-ort__profile-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
databases = "^0.5.5"
asyncpg = "^0.25.0"
pydantic = {extras = ["email"], version = "^1.9.1"}
environs = "^9.5.0"
psycopg2-binary = "^2.9.3"
passlib = "^1.7.4"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
